+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support websockets (via `--broker ws://` or `--broker wss://`)
 - Mouse clicks now select the overview / JSON Payload area
 - Home/End key support for overview and JSON Payload area
+- PageUp/Down key support for the overview
 
 ### Changed